# public/json/swagger-api.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_index","summary":"首页","parameters":[],"responses":{"200":{"description":""}},"tags":["APP主模块"]}},"/capital/login":{"post":{"operationId":"CapitalController_userLogin","summary":"用户登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":""}},"tags":["系统"]}},"/capital/role_route":{"get":{"operationId":"CapitalController_roleMenu","summary":"路由权限获取管理系统菜单列表","parameters":[{"name":"roleCode","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["系统"],"security":[{"jwt":[]}]}},"/capital/user/save":{"post":{"operationId":"UserController_create","summary":"创建用户","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]}},"/capital/user/find_page":{"post":{"operationId":"UserController_findPage","summary":"条件并分页获取用户列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageUserDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]}},"/capital/user/admin_info":{"get":{"operationId":"UserController_findInfo","summary":"获取用户信息","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]}},"/capital/user/update_role_code":{"put":{"operationId":"UserController_updateUserRole","summary":"修改用户权限","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleCodeUserDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]}},"/capital/user/update_user_info":{"put":{"operationId":"UserController_updateUserInfo","summary":"修改用户基本信息","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserInfoDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]}},"/capital/user/{userId}":{"delete":{"operationId":"UserController_remove","summary":"删除用户","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]}},"/capital/role/role_page":{"post":{"operationId":"RoleController_getRolePage","summary":"条件并分页获取权限列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["权限"],"security":[{"jwt":[]}]}},"/capital/role/all":{"get":{"operationId":"RoleController_findAll","summary":"获取全部权限列表","parameters":[],"responses":{"200":{"description":""}},"tags":["权限"],"security":[{"jwt":[]}]}},"/capital/role/save":{"post":{"operationId":"RoleController_save","summary":"新增权限","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["权限"],"security":[{"jwt":[]}]}},"/capital/role/update":{"put":{"operationId":"RoleController_update","summary":"修改权限列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["权限"],"security":[{"jwt":[]}]}},"/capital/role/{roleId}":{"delete":{"operationId":"RoleController_remove","summary":"删除权限","parameters":[{"name":"roleId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["权限"],"security":[{"jwt":[]}]}},"/capital/role/api_all":{"get":{"operationId":"RoleController_findApiAll","summary":"获取全部api接口列表","parameters":[],"responses":{"200":{"description":""}},"tags":["权限"],"security":[{"jwt":[]}]}},"/capital/menu/save":{"post":{"operationId":"MenuController_save","summary":"新增系统菜单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"201":{"description":""}},"tags":["菜单"],"security":[{"jwt":[]}]}},"/capital/menu/find_term":{"get":{"operationId":"MenuController_findAll","summary":"条件查询获取管理系统全部菜单列表","parameters":[{"name":"name","required":false,"in":"query","description":"菜单的唯一标识","schema":{"type":"string"}},{"name":"title","required":false,"in":"query","description":"菜单名称","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["菜单"],"security":[{"jwt":[]}]}},"/capital/menu/update":{"put":{"operationId":"MenuController_update","summary":"修改系统菜单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMenuDto"}}}},"responses":{"200":{"description":""}},"tags":["菜单"],"security":[{"jwt":[]}]}},"/capital/menu/{menuId}":{"delete":{"operationId":"MenuController_remove","summary":"删除系统菜单","parameters":[{"name":"menuId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["菜单"],"security":[{"jwt":[]}]}},"/file":{"get":{"operationId":"FileController_index","summary":"文件内容","parameters":[],"responses":{"200":{"description":""}},"tags":["文件"]}},"/file/image/upload":{"post":{"operationId":"ImageController_uploadImage","summary":"单图片上传，接收 multipart/form-data 的数据","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadImageDto"}}}},"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/public":{"get":{"operationId":"ImageController_getPublic","summary":"获取图片目录","parameters":[],"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/only_public":{"get":{"operationId":"ImageController_getOnlyPublic","summary":"查询只有图片文件没有数据的文件","parameters":[],"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/not_use":{"get":{"operationId":"ImageController_getOntUse","summary":"查询未使用的图片","parameters":[],"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/find_all":{"get":{"operationId":"ImageController_findAll","summary":"获取图片全部列表数据","parameters":[],"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/find_page":{"post":{"operationId":"ImageController_findPage","summary":"条件并分页获取图片数据列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageImageDto"}}}},"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/remove_public/{fileName}":{"delete":{"operationId":"ImageController_removePublic","summary":"删除图片目录下的图片","parameters":[{"name":"fileName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/remove_public_all":{"delete":{"operationId":"ImageController_removePublicAll","summary":"批量删除图片目录下的图片","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemovePublicAllImageDto"}}}},"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/remove_data/{imageId}":{"delete":{"operationId":"ImageController_removeDataAll","summary":"批量删除图片下的数据","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveDataAllImageDto"}}}},"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/remove_public_data/{imageId}":{"delete":{"operationId":"ImageController_removePublicAndData","summary":"删除 图片目录下的图片 和 图片下的数据","parameters":[{"name":"imageId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["图片"]}},"/file/image/remove_public_data_all":{"delete":{"operationId":"ImageController_removePublicAndDataAll","summary":"批量删除 图片目录下的图片 和 图片下的数据","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemovePublicAndDataAllImageDto"}}}},"responses":{"200":{"description":""}},"tags":["图片"]}},"/money":{"get":{"operationId":"MoneyController_index","summary":"金钱内容","parameters":[],"responses":{"200":{"description":""}},"tags":["金钱"]}},"/money/we-chat/upload":{"post":{"operationId":"WeChatController_upload","summary":"微信账单导入","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadWeChatDto"}}}},"responses":{"200":{"description":""}},"tags":["微信"],"security":[{"jwt":[]}]}},"/money/we-chat/save":{"post":{"operationId":"WeChatController_save","summary":"创建微信账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWeChatDto"}}}},"responses":{"200":{"description":""}},"tags":["微信"],"security":[{"jwt":[]}]}},"/money/we-chat/batch-save":{"post":{"operationId":"WeChatController_batchSave","summary":"批量创建微信账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWeChatBatchDto"}}}},"responses":{"200":{"description":""}},"tags":["微信"],"security":[{"jwt":[]}]}},"/money/we-chat/find_page":{"post":{"operationId":"WeChatController_findPage","summary":"条件并分页获取微信账单列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageWeChatDto"}}}},"responses":{"200":{"description":""}},"tags":["微信"],"security":[{"jwt":[]}]}},"/money/we-chat/update":{"put":{"operationId":"WeChatController_update","summary":"修改微信账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWeChatDto"}}}},"responses":{"200":{"description":""}},"tags":["微信"],"security":[{"jwt":[]}]}},"/money/ali-pay/upload":{"post":{"operationId":"AliPayController_upload","summary":"支付宝账单导入","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadAliPayDto"}}}},"responses":{"200":{"description":""}},"tags":["支付宝"],"security":[{"jwt":[]}]}},"/money/ali-pay/batch-save":{"post":{"operationId":"AliPayController_batchSave","summary":"批量创建支付宝账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAliPayBatchDto"}}}},"responses":{"200":{"description":""}},"tags":["支付宝"],"security":[{"jwt":[]}]}},"/money/ali-pay/find_page":{"post":{"operationId":"AliPayController_findPage","summary":"条件并分页获取支付宝账单列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageAliPayDto"}}}},"responses":{"200":{"description":""}},"tags":["支付宝"],"security":[{"jwt":[]}]}},"/money/ali-pay/update":{"put":{"operationId":"AliPayController_update","summary":"修改支付宝账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAliPayDto"}}}},"responses":{"200":{"description":""}},"tags":["支付宝"],"security":[{"jwt":[]}]}},"/money/bank/upload":{"post":{"operationId":"BankController_upload","summary":"银行账单导入","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadBankDto"}}}},"responses":{"200":{"description":""}},"tags":["银行"],"security":[{"jwt":[]}]}},"/money/bank/batch-save":{"post":{"operationId":"BankController_batchSave","summary":"批量创建银行账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBankBatchDto"}}}},"responses":{"200":{"description":""}},"tags":["银行"],"security":[{"jwt":[]}]}},"/money/bank/find_page":{"post":{"operationId":"BankController_findPage","summary":"条件并分页获取银行账单列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageBankDto"}}}},"responses":{"200":{"description":""}},"tags":["银行"],"security":[{"jwt":[]}]}},"/money/bank/update":{"put":{"operationId":"BankController_update","summary":"修改银行账单","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBankDto"}}}},"responses":{"200":{"description":""}},"tags":["银行"],"security":[{"jwt":[]}]}},"/blog":{"get":{"operationId":"BlogController_index","summary":"博客内容","parameters":[],"responses":{"200":{"description":""}},"tags":["博客"]}},"/blog/article/find_page":{"post":{"operationId":"ArticleController_findPage","summary":"条件并分页获取文章列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageArticleDto"}}}},"responses":{"200":{"description":""}},"tags":["文章"]}},"/blog/article/save":{"post":{"operationId":"ArticleController_save","summary":"新增文章","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleDto"}}}},"responses":{"200":{"description":""}},"tags":["文章"],"security":[{"jwt":[]}]}},"/blog/article/update":{"put":{"operationId":"ArticleController_update","summary":"修改文章","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArticleDto"}}}},"responses":{"200":{"description":""}},"tags":["文章"],"security":[{"jwt":[]}]}},"/blog/article/{articleId}":{"delete":{"operationId":"ArticleController_remove","summary":"删除文章","parameters":[{"name":"articleId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章"],"security":[{"jwt":[]}]}},"/blog/article-category/save":{"post":{"operationId":"ArticleCategoryController_save","summary":"新增文章分类","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["文章分类"],"security":[{"jwt":[]}]}},"/blog/article-category/all":{"get":{"operationId":"ArticleCategoryController_findAll","summary":"获取全部文章分类列表","parameters":[],"responses":{"200":{"description":""}},"tags":["文章分类"]}},"/blog/article-category/{articleCategoryId}":{"delete":{"operationId":"ArticleCategoryController_remove","summary":"删除文章分类","parameters":[{"name":"articleCategoryId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章分类"],"security":[{"jwt":[]}]}}},"info":{"title":"NestJs博客API","description":"我的测试博客API \n\n swagger的JSON文件：/public/json/swagger-api.json","version":"1.1.3","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"securitySchemes":{"jwt":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名","example":"nest"},"password":{"type":"string","description":"密码","example":"123456"}},"required":["username","password"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名","example":"text"},"password":{"type":"string","description":"密码","example":"123123"},"nickname":{"type":"string","description":"昵称"},"avatar":{"type":"string","description":"用户头像"},"roleCode":{"type":"string","description":"角色标识"}},"required":["username","password","nickname","avatar","roleCode"]},"PageUserDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"nickname":{"type":"string","description":"昵称","default":""},"username":{"type":"string","description":"用户名","default":""}},"required":["size","current"]},"UpdateRoleCodeUserDto":{"type":"object","properties":{"userId":{"type":"string","description":"用户id"},"roleCode":{"type":"string","description":"角色标识"}},"required":["userId","roleCode"]},"UpdateUserInfoDto":{"type":"object","properties":{"nickname":{"type":"string","description":"昵称"},"avatar":{"type":"string","description":"用户头像"},"username":{"type":"string","description":"账号名"}},"required":["nickname","avatar","username"]},"PageRoleDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"name":{"type":"string","description":"角色名称","default":""},"roleCode":{"type":"string","description":"角色标识","default":""}},"required":["size","current"]},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string","description":"角色名称"},"roleCode":{"type":"string","description":"角色标识"},"roleType":{"type":"number","description":"角色权限类型"},"menuPermission":{"description":"菜单权限","type":"array","items":{"type":"string"}},"apiPermission":{"description":"api权限","type":"array","items":{"type":"string"}}},"required":["name","roleCode","roleType","menuPermission","apiPermission"]},"UpdateRoleDto":{"type":"object","properties":{"name":{"type":"string","description":"角色名称"},"roleCode":{"type":"string","description":"角色标识"},"roleType":{"type":"number","description":"角色权限类型"},"menuPermission":{"description":"菜单权限","type":"array","items":{"type":"string"}},"apiPermission":{"description":"api权限","type":"array","items":{"type":"string"}},"roleId":{"type":"string","description":"权限id"}},"required":["roleId"]},"CreateMenuDto":{"type":"object","properties":{"name":{"type":"string","description":"菜单的唯一标识"}},"required":["name"]},"UpdateMenuDto":{"type":"object","properties":{"name":{"type":"string","description":"菜单的唯一标识"},"menuId":{"type":"string","description":"菜单id"}},"required":["menuId"]},"UploadImageDto":{"type":"object","properties":{"image":{"type":"string","description":"单图片上传","format":"binary"}},"required":["image"]},"PageImageDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"name":{"type":"string","description":"图片名称","default":""}},"required":["size","current"]},"RemovePublicAllImageDto":{"type":"object","properties":{"fileNameArr":{"type":"string[]","description":"文件全称数组"}},"required":["fileNameArr"]},"RemoveDataAllImageDto":{"type":"object","properties":{"imageIdArr":{"type":"string[]","description":"图片id数组"}},"required":["imageIdArr"]},"RemovePublicAndDataAllImageDto":{"type":"object","properties":{"imageIdArr":{"type":"string[]","description":"图片id数组"}},"required":["imageIdArr"]},"UploadWeChatDto":{"type":"object","properties":{"file":{"type":"string","description":"单文件上传","format":"binary"}},"required":["file"]},"CreateWeChatDto":{"type":"object","properties":{"tradeTime":{"type":"string","description":"交易时间"},"tradeType":{"type":"string","description":"交易类型"},"tradeOtherPerson":{"type":"string","description":"交易对方"},"tradeOtherPersonRemarks":{"type":"string","description":"交易对方备注"},"goods":{"type":"string","description":"商品"},"incomeOrPay":{"type":"string","description":"收/支"},"moneyAmount":{"type":"number","description":"金额(元)"},"paymentMethod":{"type":"string","description":"支付方式"},"currentStatus":{"type":"string","description":"当前状态"},"remarks":{"type":"string","description":"备注"},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"explain":{"type":"string","description":"账单说明"},"place":{"type":"string","description":"使用地点"},"billType":{"type":"number","description":"账单类型"}},"required":["tradeTime","tradeType","tradeOtherPerson","tradeOtherPersonRemarks","goods","incomeOrPay","moneyAmount","paymentMethod","currentStatus","remarks","inflowOrOutflow","explain","place","billType"]},"CreateWeChatBatchDto":{"type":"object","properties":{"batches":{"description":"批量导入","type":"array","items":{"type":"string"}}},"required":["batches"]},"PageWeChatDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"tradeOtherPerson":{"type":"string","description":"交易对方","default":""},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"billType":{"type":"number","description":"账单类型"}},"required":["size","current"]},"UpdateWeChatDto":{"type":"object","properties":{"weChatId":{"type":"string","description":"微信账单id"},"tradeOtherPersonRemarks":{"type":"string","description":"交易对方备注"},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"explain":{"type":"string","description":"账单说明"},"place":{"type":"string","description":"使用地点"},"billType":{"type":"number","description":"账单类型"}},"required":["weChatId","tradeOtherPersonRemarks","inflowOrOutflow","explain","place","billType"]},"UploadAliPayDto":{"type":"object","properties":{"file":{"type":"string","description":"单文件上传","format":"binary"}},"required":["file"]},"CreateAliPayBatchDto":{"type":"object","properties":{"batches":{"description":"批量导入","type":"array","items":{"type":"string"}}},"required":["batches"]},"PageAliPayDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"tradeOtherPerson":{"type":"string","description":"交易对方","default":""},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"billType":{"type":"number","description":"账单类型"}},"required":["size","current"]},"UpdateAliPayDto":{"type":"object","properties":{"aliPayId":{"type":"string","description":"支付宝账单id"},"tradeOtherPersonRemarks":{"type":"string","description":"交易对方备注"},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"explain":{"type":"string","description":"账单说明"},"place":{"type":"string","description":"使用地点"},"billType":{"type":"number","description":"账单类型"}},"required":["aliPayId","tradeOtherPersonRemarks","inflowOrOutflow","explain","place","billType"]},"UploadBankDto":{"type":"object","properties":{"file":{"type":"string","description":"单文件上传","format":"binary"}},"required":["file"]},"CreateBankBatchDto":{"type":"object","properties":{"batches":{"description":"批量导入","type":"array","items":{"type":"string"}}},"required":["batches"]},"PageBankDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"tradeOtherPerson":{"type":"string","description":"交易对方","default":""},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"bankBillType":{"type":"number","description":"银行账单类型"}},"required":["size","current"]},"UpdateBankDto":{"type":"object","properties":{"bankId":{"type":"string","description":"银行账单id"},"tradeOtherPersonRemarks":{"type":"string","description":"交易对方备注"},"inflowOrOutflow":{"type":"number","description":"流入/流出"},"explain":{"type":"string","description":"账单说明"},"place":{"type":"string","description":"使用地点"},"otherCost":{"type":"number","description":"其它费用"},"bankBillType":{"type":"number","description":"银行账单类型"}},"required":["bankId","tradeOtherPersonRemarks","inflowOrOutflow","explain","place","otherCost","bankBillType"]},"PageArticleDto":{"type":"object","properties":{"size":{"type":"number","description":"每页数量","example":10},"current":{"type":"number","description":"当前页","example":1},"title":{"type":"string","description":"文章标题","default":""},"categoryVal":{"type":"number","description":"文章分类"}},"required":["size","current"]},"CreateArticleDto":{"type":"object","properties":{"title":{"type":"string","description":"文章的标题","example":"文章的标题"},"brief":{"type":"string","description":"文章的简介","example":"文章的简介"},"htmlContent":{"type":"string","description":"文章的html内容","example":""},"markdownContent":{"type":"string","description":"文章的markdown内容","example":""},"categoryVal":{"type":"number","description":"文章的类型标识","example":1}},"required":["title","brief","htmlContent","markdownContent","categoryVal"]},"UpdateArticleDto":{"type":"object","properties":{"title":{"type":"string","description":"文章的标题","example":"文章的标题"},"brief":{"type":"string","description":"文章的简介","example":"文章的简介"},"htmlContent":{"type":"string","description":"文章的html内容","example":""},"markdownContent":{"type":"string","description":"文章的markdown内容","example":""},"categoryVal":{"type":"number","description":"文章的类型标识","example":1},"articleId":{"type":"string","description":"文章id"}},"required":["articleId"]},"CreateArticleCategoryDto":{"type":"object","properties":{"name":{"type":"string","description":"文章分类名称","example":"测试"}},"required":["name"]}}}}